requirements.txt
Django==2.0
psycopg2
django-ckeditor==5.4.0
django-braces==1.12.0
django-widget-tweaks-1.4.2
Pillow==4.3.0
awesome-slugify==1.6.5
telepot==12.5
psycopg2-binary==2.7.4
opbeat==3.6.1
